Switch to 128-bit token ids
SHA2 has 128 bits of security anyway, so the 256 bits of randomness were
not fully used.

As a result, the size of cookie information will go from 86 to 44.

This is backwards-compatible: old, bigger cookie tokens are still read
properly. The on-disk hash of old and new secret obviously has the same
size, and the filenames specifying session ids are smaller, which is not
an issue.